kamien_milowy_3/web/requirements.txt
Flask
Flask-Session
gunicorn
python-dotenv
redis
bcrypt
PyJWT
requests
